@@ -9,35 +9,42 @@
 		}
 	],
 	"require": {
-		"php": ">=7.4",
+		"php": ">=8.4",
 		"monolog/monolog": "^2.0",
-		"myclabs/php-enum": "^1.7",
-		"psr/container": "1.0.0"
+		"psr/container": "^2.0.2"
 	},
 	"require-dev": {
-		"holonet/hdev": "~1.0.0",
-		"phpunit/phpunit": "^8.4.1"
+		"holonet/hdev": "~1.1.0@dev",
+		"phpunit/phpunit": "^10.0.0",
+		"spatie/phpunit-snapshot-assertions": "dev-main"
 	},
 	"provide": {
-		"psr/container-implementation": "1.0.0"
+		"psr/container-implementation": "^2.0.2"
 	},
 	"repositories": [
 		{
 			"type": "composer",
-			"url": "https://holonet.easylabs.ch/hgit/composer/"
+			"url": "https://holonetz.synology.me/hgit/composer/"
 		}
 	],
+	"minimum-stability": "dev",
 	"autoload": {
 		"psr-4": {
 			"holonet\\common\\": "src/"
 		},
 		"files": [
-			"src/functions.php"
+			"src/functions.php",
+			"src/kvm_functions.php"
 		]
 	},
 	"autoload-dev": {
 		"psr-4": {
 			"holonet\\common\\tests\\": "tests/"
+		}
+	},
+	"config": {
+		"allow-plugins": {
+			"ergebnis/composer-normalize": true
 		}
 	},
 	"extra": {
@@ -48,14 +55,15 @@
 	},
 	"scripts": {
 		"fix": [
-			"@composer normalize --diff",
-			"@php -d memory_limit=-1 vendor/friendsofphp/php-cs-fixer/php-cs-fixer fix --verbose --ansi"
+			"@composer normalize --diff --ansi",
+			"@php -d memory_limit=-1 vendor/friendsofphp/php-cs-fixer/php-cs-fixer fix --verbose --ansi --diff"
 		],
 		"test": [
-			"@composer validate",
-			"@composer normalize --dry-run --diff",
-			"@php -d memory_limit=-1 vendor/phpunit/phpunit/phpunit",
-			"@php -d memory_limit=-1 vendor/friendsofphp/php-cs-fixer/php-cs-fixer fix --verbose --dry-run --ansi",
+			"@composer validate --ansi",
+			"@composer normalize --dry-run --diff --ansi",
+			"@php vendor/bin/hdev sync-dev-stubs --no-interactive",
+			"@php -d xdebug.mode=coverage -d memory_limit=-1 vendor/phpunit/phpunit/phpunit -c phpunit.xml --colors=always",
+			"@php -d memory_limit=-1 vendor/friendsofphp/php-cs-fixer/php-cs-fixer fix --verbose --dry-run --ansi --diff",
 			"@php -d memory_limit=-1 vendor/vimeo/psalm/psalm"
 		]
 	}
