@@ -38,7 +38,7 @@ class TokeniserClassDiscovery extends ClassDiscovery {
 			for (; $i < count($tokens); $i++) {
 				if ($tokens[$i][0] === \T_NAMESPACE) {
 					for ($j = $i + 1; $j < count($tokens); $j++) {
-						if ($tokens[$j][0] === \T_STRING) {
+						if ($tokens[$j][0] === \T_STRING || $tokens[$j][0] === \T_NAME_QUALIFIED) {
 							$namespace .= '\\'.$tokens[$j][1];
 						} elseif ($tokens[$j] === '{' || $tokens[$j] === ';') {
 							break;
