@@ -47,12 +47,11 @@ class ErrorDispatcher {
 	}
 
 	/**
-	 * registers the our callbacks as error handlers/exception handlers/shutdown function with the SPL.
+	 * registers our callbacks as error handlers/exception handlers/shutdown function with the SPL.
 	 */
-	public function register(): void {
+	public function __construct() {
 		/**
 		 * @psalm-suppress InvalidArgument
-		 * @psalm-suppress MissingClosureParamType
 		 */
 		set_error_handler(function (...$args): void {
 			foreach ($this->errorHandlers as $handler) {
@@ -64,10 +63,6 @@ class ErrorDispatcher {
 			}
 		});
 
-		/**
-		 * @psalm-suppress InvalidArgument
-		 * @psalm-suppress MissingClosureParamType
-		 */
 		set_exception_handler(function (...$args): void {
 			foreach ($this->exceptionHandlers as $handler) {
 				$handler(...$args);
@@ -79,7 +74,6 @@ class ErrorDispatcher {
 		});
 
 		/**
-		 * @psalm-suppress InvalidArgument
 		 * @psalm-suppress MissingClosureParamType
 		 */
 		register_shutdown_function(function (...$args): void {
